Do not delete physical hardpoints on systems
TF state changes can eliminate awareness of resouce elements which
then causes the provider to Delete them; a logical absurdity given
that these are physical components in physical systems and TF's
state describing them does not actually remove hardware from the
machine as MaaS is being told to believe.

Update block_device to remove partitions but leave the actual disk
visible to MaaS for subsequent action without having to commission
the machine again.

Update network_interface_physical to Disconnect instead of Delete
so as to avoid removing potentially relevant attributes while not
allowing it to logically interfere with operations.