Optimize getMachine function
The getMachine is incredibly inefficient fetching a list of all
machines and their components in order to look up the name or ID of
just one and then doing it all over again for each component of
that machine's configuration. The API provides for a Machine.Get
call when using the machine ID or a params struct for Machines.Get
which can specify Hostname - a commonly used element in resource
calls.

Return a machine found by its ID before performing lookup on the
list of all machines, then try to find a machine in the list by
the hostname parameter, and finally fall back to the current list
lookup dumping all data and iterating through it for a few fields.

Testing:
  Qualified on lab cluster showing reduction of 30m plan to < 3m.