feat: 230 extend tagging to support nics (#290)
An additional resource to manage the tags of network interfaces. This resource must not be used in conjunction with the nested tag keys in resources such as `maas_network_interface_physical`. Other examples of additional supporting resources that follow this design pattern can be found in the [AWS provider](https://registry.terraform.io/providers/hashicorp/aws/latest/docs/guides/resource-tagging#ignoring-changes-to-specific-tags).