fix: getMachine defaults to system-id
When trying to resolve an identifier into a Machine, first do a simple
one-machine query to lookup the identifier by SystemID. Only once that
fails do we fall back to the previous algorithm, which would read all
machines, and do a one-by-one comparison to see if that was a SystemID,
a Hostname, a FQDN, or the boot MACAddress.

A very common pattern with Terraform is going to be defining a Machine
resource, and then referencing it later using `mymachine.id` and so we
can do a single lookup on that attribute, rather than reading all
machines for any reference.

fixes: #343